+++ translator/src/main/scala/se/umu/cs/rai/scopus/translator/Main.scala
@@ -1,7 +1,9 @@
 package se.umu.cs.rai.scopus.translator
 
+import java.io.File
 import java.io.FileWriter
-import java.nio.file.{Files, Paths}
+import java.nio.file.Files
+import java.nio.file.Paths
 
 object EntryPoint {
   def main(args: Array[String]): Unit = Main().main(args)
@@ -24,6 +26,7 @@ case class Main() {
     "\nwhere" +
     "\n" +
     "\n  SCOPUS_SCALA_INPUT is used to create the Scopus input file and Scala output file. " +
+    "\nIf it is a directory, it scans recursively the directory to translate Scopus files." +
     "\nIf the extension is " + ScopusExtension + ", the output file has extension " + ScalaExtension + "." +
     "\nOtherwise, the extension " + ScopusExtension + " and " + ScalaExtension + " are appended to create the input and output files respectively." +
     "\n" +
@@ -43,19 +46,18 @@ case class Main() {
   }
 
   def main(args: Array[String]): Unit =
-    if ( args.length == 1
-    ) {
-      val (inputFileName, outputFileName) = getInputOutputFileNames(args(0))
-      translate(inputFileName, outputFileName)
-    }
-    else if ( args.length == 2
-    ) {
-      val inputFileName = args(0)
-      val outputFileName = args(1)
-      translate(inputFileName, outputFileName)
-    }
+    if ( args.length == 1 ) process_directory(args(0))
+    else if ( args.length == 2 ) translate(args(0), args(1))
     else println(getTitleAndVersion + Help)
 
+  def process_directory(start: String) =
+    DirectoryScanner()
+      .get_scopus_files(new File(start))
+      .map(file => {
+        val file_name = file.getAbsolutePath
+        val (inputFileName, outputFileName) = getInputOutputFileNames(file_name)
+        translate(inputFileName, outputFileName)
+      })
 
   def getInputOutputFileNames(inputName: String): (String, String) =
     if ( inputName.endsWith(ScopusExtension)